requirements.txt
aniso8601==9.0.1
anyio==4.3.0
astroid==2.1.0
autopep8==1.4.3
boto3==1.34.57
botocore==1.34.57
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.1
coverage==7.4.3
cryptography==42.0.5
docutils==0.14
exceptiongroup==1.2.0
fastapi==0.110.0
graphene==3.2.1
graphql-core==3.2.0
graphql-relay==3.2.0
gunicorn==20.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
iniconfig==2.0.0
isort==4.3.4
Jinja2==3.1.3
jmespath==0.9.3
lazy-object-proxy==1.4.1
MarkupSafe==2.1.5
mccabe==0.6.1
mock==4.0.3
moto==4.2.14
packaging==23.2
pluggy==1.4.0
pycodestyle==2.4.0
pycparser==2.21
pydantic==1.10.14
pylint==2.2.2
pynamodb==6.0.0
pytest==8.0.2
python-dateutil==2.7.5
pytz==2018.7
PyYAML==6.0.1
requests==2.31.0
responses==0.25.0
rfc3986==1.5.0
Rx==1.6.1
s3transfer==0.10.0
singledispatch==3.4.0.3
six==1.12.0
sniffio==1.3.1
starlette==0.36.3
starlette-graphene3==0.6.0
tomli==2.0.1
typing==3.6.6
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.20.0
Werkzeug==3.0.1
wrapt==1.10.11
xmltodict==0.13.0
